Compilation fix
